sort routes into individual routers
