+++ backend/src/utils.ts
@@ -1,3 +1,6 @@
+import { Response as ExpressResponse } from "express";
+import { User } from "./domain/User.js";
+
 export function isProd() {
     return process.env.NODE_ENV === "production";
 }
@@ -12,4 +15,11 @@ export function buildUrl(path: string) {
     } else {
         return "http://localhost:5001" + path;
     }
+}
+interface Locals extends Record<string, any> {
+    user?: User;
+}
+
+export interface Response extends ExpressResponse {
+    locals: Locals;
 }
